add TableXPositionContext
